Convert one-hot tensor to float to be able to use it with cross entropy.

+++ ssi/machine_learning/pytorch.py
@@ -78,7 +78,7 @@ class ParquetDataset(torch.utils.data.Dataset):
         label_tensor = torch.tensor(
             self.label_encoder.transform([sample[self.target_column]]), dtype=torch.long)
 
-        return feature_tensor, F.one_hot(label_tensor, num_classes=len(self.label_encoder.classes_))
+        return feature_tensor, F.one_hot(label_tensor, num_classes=len(self.label_encoder.classes_)).float()
 
 
 class TorchLogisticRegression(nn.Module):
